Make the code more modular
Add maven module
