+++ dev-tools/build_release.py
@@ -27,9 +27,9 @@ from release.logger import logger
 from release.git import github
 from release.git import git
 from release.announcement import announcement
-from release.utils import strings
 from release.utils import run
 from release.documentation import doc_update
+from release.maven import mvn
 
 """
  This tool builds a release from the a given elasticsearch plugin branch.
@@ -87,42 +87,10 @@ FAIL = '\033[91m'
 
 ##########################################################
 #
-# Clean logs and check JAVA and Maven
+# Clean logs
 #
 ##########################################################
-try:
-    logger.purge_log()
-    JAVA_HOME = env['JAVA_HOME']
-except KeyError:
-    raise RuntimeError("""
-  Please set JAVA_HOME in the env before running release tool
-  On OSX use: export JAVA_HOME=`/usr/libexec/java_home -v '1.7*'`""")
-
-try:
-    MVN = 'mvn'
-    # make sure mvn3 is used if mvn3 is available
-    # some systems use maven 2 as default
-    run.run('mvn3 --version', quiet=True)
-    MVN = 'mvn3'
-except RuntimeError:
-    pass
-
-
-def java_exe():
-    path = JAVA_HOME
-    return 'export JAVA_HOME="%s" PATH="%s/bin:$PATH" JAVACMD="%s/bin/java"' % (path, path, path)
-
-
-def verify_java_version(version):
-    s = os.popen('%s; java -version 2>&1' % java_exe()).read()
-    if ' version "%s.' % version not in s:
-        raise RuntimeError('got wrong version for java %s:\n%s' % (version, s))
-
-
-def verify_mvn_java_version(version, mvn):
-    s = os.popen('%s; %s --version 2>&1' % (java_exe(), mvn)).read()
-    if 'Java version: %s' % version not in s:
-        raise RuntimeError('got wrong java version for %s %s:\n%s' % (mvn, version, s))
+logger.purge_log()
 
 
 ##########################################################
@@ -130,51 +98,6 @@ def verify_mvn_java_version(version, mvn):
 # String and file manipulation utils
 #
 ##########################################################
-# Guess the next snapshot version number (increment last digit)
-def guess_snapshot(version):
-    digits = strings.split_version_to_digits(version)
-    source = '%s.%s.%s' % (digits[0], digits[1], digits[2])
-    destination = '%s.%s.%s' % (digits[0], digits[1], digits[2] + 1)
-    return version.replace(source, destination)
-
-
-# Checks the pom.xml for the release version. <version>2.0.0-SNAPSHOT</version>
-# This method fails if the pom file has no SNAPSHOT version set ie.
-# if the version is already on a release version we fail.
-# Returns the next version string ie. 0.90.7
-def find_release_version(src_branch):
-    git.checkout(src_branch)
-    with open(POM_FILE, encoding='utf-8') as file:
-        for line in file:
-            match = re.search(r'<version>(.+)-SNAPSHOT</version>', line)
-            if match:
-                return match.group(1)
-        raise RuntimeError('Could not find release version in branch %s' % src_branch)
-
-
-# extract a value from pom.xml after a given line
-def find_from_pom(tag, first_line=None):
-    with open(POM_FILE, encoding='utf-8') as file:
-        previous_line_matched = False
-        if first_line is None:
-            previous_line_matched = True
-        for line in file:
-            if previous_line_matched:
-                match = re.search(r'<%s>(.+)</%s>' % (tag, tag), line)
-                if match:
-                    return match.group(1)
-
-            if first_line is not None:
-                match = re.search(r'%s' % first_line, line)
-                if match:
-                    previous_line_matched = True
-
-        if first_line is not None:
-            raise RuntimeError('Could not find %s in pom.xml file after %s' % (tag, first_line))
-        else:
-            raise RuntimeError('Could not find %s in pom.xml file' % tag)
-
-
 # Get artifacts which have been generated in target/releases
 def get_artifacts(artifact_id, release):
     artifact_path = ROOT_DIR + '/target/releases/%s-%s.zip' % (artifact_id, release)
@@ -198,30 +121,6 @@ def generate_checksums(release_file):
     res += [os.path.join(directory, checksum_file), release_file]
     return res
 
-
-
-##########################################################
-#
-# Maven commands
-#
-##########################################################
-# Run a given maven command
-def run_mvn(*cmd):
-    for c in cmd:
-        run.run('%s; %s -f %s %s' % (java_exe(), MVN, POM_FILE, c))
-
-
-# Run deploy or package depending on dry_run
-# Default to run mvn package
-# When run_tests=True a first mvn clean test is run
-def build_release(run_tests=False, dry_run=True):
-    target = 'deploy'
-    tests = '-DskipTests'
-    if run_tests:
-        tests = ''
-    if dry_run:
-        target = 'package'
-    run_mvn('clean %s %s' % (target, tests))
 
 
 ##########################################################
@@ -338,8 +237,8 @@ def check_environment_and_commandline_tools():
     command = 'python %s/upload-s3.py' % location
     run_and_print('Testing boto python dependency...   ', partial(check_command_exists, 'python-boto', command))
 
-    run_and_print('Checking java version...            ', partial(verify_java_version, '1.7'))
-    run_and_print('Checking java mvn version...        ', partial(verify_mvn_java_version, '1.7', MVN))
+    run_and_print('Checking java version...            ', partial(mvn.verify_java_version, '1.7'))
+    run_and_print('Checking java mvn version...        ', partial(mvn.verify_mvn_java_version, '1.7', mvn.MVN))
 
 
 if __name__ == '__main__':
@@ -389,21 +288,22 @@ if __name__ == '__main__':
 
     print(''.join(['-' for _ in range(80)]))
     print('Preparing Release from branch [%s] running tests: [%s] dryrun: [%s]' % (src_branch, run_tests, dry_run))
-    print('  JAVA_HOME is [%s]' % JAVA_HOME)
-    print('  Running with maven command: [%s] ' % (MVN))
+    print('  JAVA_HOME is [%s]' % mvn.JAVA_HOME)
+    print('  Running with maven command: [%s] ' % mvn.MVN)
 
-    release_version = find_release_version(src_branch)
-    artifact_id = find_from_pom('artifactId')
-    artifact_name = find_from_pom('name')
-    artifact_description = find_from_pom('description')
-    project_url = find_from_pom('url')
+    git.checkout(src_branch)
+    release_version = mvn.find_release_version(src_branch)
+    artifact_id = mvn.find_from_pom('artifactId')
+    artifact_name = mvn.find_from_pom('name')
+    artifact_description = mvn.find_from_pom('description')
+    project_url = mvn.find_from_pom('url')
 
     try:
-        elasticsearch_version = find_from_pom('elasticsearch.version')
+        elasticsearch_version = mvn.find_from_pom('elasticsearch.version')
     except RuntimeError:
         # With projects using elasticsearch-parent project, we need to consider elasticsearch version
         # to be after <artifactId>elasticsearch-parent</artifactId>
-        elasticsearch_version = find_from_pom('version', '<artifactId>elasticsearch-parent</artifactId>')
+        elasticsearch_version = mvn.find_from_pom('version', '<artifactId>elasticsearch-parent</artifactId>')
 
     print('  Artifact Id: [%s]' % artifact_id)
     print('  Release version: [%s]' % release_version)
@@ -412,7 +312,7 @@ if __name__ == '__main__':
         raise RuntimeError('Can not release with a SNAPSHOT elasticsearch dependency: %s' % elasticsearch_version)
 
     # extract snapshot
-    default_snapshot_version = guess_snapshot(release_version)
+    default_snapshot_version = mvn.guess_snapshot(release_version)
     snapshot_version = input('Enter next snapshot version [%s]:' % default_snapshot_version)
     snapshot_version = snapshot_version or default_snapshot_version
 
@@ -429,7 +329,7 @@ if __name__ == '__main__':
         master_hash = git.get_head_hash()
         git.checkout(src_branch)
         version_hash = git.get_head_hash()
-        run_mvn('clean')  # clean the env!
+        mvn.clean()  # clean the env!
         git.create_release_branch(remote, 'master', release_version)
         print('  Created release branch [%s]' % (git.release_branch('master', release_version)))
         git.create_release_branch(remote, src_branch, release_version)
@@ -460,7 +360,7 @@ if __name__ == '__main__':
             print('  Running maven builds now and publish to sonatype - run-tests [%s]' % run_tests)
         else:
             print('  Running maven builds now run-tests [%s]' % run_tests)
-        build_release(run_tests=run_tests, dry_run=dry_run)
+        mvn.build_release(run_tests=run_tests, dry_run=dry_run)
         artifact = get_artifacts(artifact_id, release_version)
         artifact_and_checksums = generate_checksums(artifact)
         print(''.join(['-' for _ in range(80)]))
@@ -468,7 +368,7 @@ if __name__ == '__main__':
         ########################################
         # Start update process in master branch
         ########################################
-        git.checkout(release_branch('master', release_version))
+        git.checkout(git.release_branch('master', release_version))
         doc_update.update_documentation_to_released_version(README_FILE, project_url, release_version, src_branch,
                                                  elasticsearch_version)
         doc_update.set_install_instructions(README_FILE, artifact_id, release_version)
@@ -532,8 +432,8 @@ Release successful pending steps:
             run.run('git tag -d v%s' % release_version)
 
         # we delete this one anyways
-        run.run('git branch -D %s' % (release_branch('master', release_version)))
-        run.run('git branch -D %s' % (release_branch(src_branch, release_version)))
+        run.run('git branch -D %s' % (git.release_branch('master', release_version)))
+        run.run('git branch -D %s' % (git.release_branch(src_branch, release_version)))
 
         # Checkout the branch we started from
         git.checkout(src_branch)
